refactor: create `AppBody`
